mempool: Hide `MempoolInterfaceImpl` behind a trait
