+++ mempool/src/interface/mempool_interface.rs
@@ -19,6 +19,7 @@ use common::{
     primitives::Id,
 };
 use std::sync::Arc;
+use subsystem::{CallRequest, ShutdownRequest};
 
 #[async_trait::async_trait]
 pub trait MempoolInterface: Send + Sync {
@@ -37,4 +38,9 @@ pub trait MempoolInterface: Send + Sync {
         &mut self,
         handler: Arc<dyn Fn(MempoolEvent) + Send + Sync>,
     ) -> Result<(), Error>;
+}
+
+#[async_trait::async_trait]
+pub trait MempoolSubsystemInterface: MempoolInterface + 'static {
+    async fn run(self, call_rq: CallRequest<dyn MempoolInterface>, shut_rq: ShutdownRequest);
 }
